reference/eai-hub-agentic-starter-v2/tools/ngbse/Legacyfiles/NL_Blindspot_Engine_v2/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
python-docx>=1.1.0
